Prevent crash on shutdown due to mpi call after finalize
The subsnap object in HBT.c goes out of scope after
MPI_Finalize and its destructor calls MPI_Type_Free.
This causes the program to abort with a non-zero exit code.

The fix is just to check if MPI has been finalized in
the destructor and if it has we don't free the type.